+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Changed
 - the AI now requires Python version 3.8 or higher
+- DroneMode renamed to Speech Optimization
 
 ## [1.1.0.5]
 ### Fixed